# docs/openapi.front.yaml
openapi: 3.1.0
info:
  title: Lianchuan Storefront API
  version: 0.1.0
  summary: 前台电商网站接口契约
  description: |
    本文档定义前台商城所需的 REST API 契约，覆盖首页、分类、搜索、商品详情、购物车、结账、地址、订单、认证、会员资料、询价、收藏等能力。
    该文档是前台与服务端联调的唯一接口真源。
servers:
  - url: http://localhost:5100
    description: Local admin API (vexmotor-admin)
  - url: https://example.com
    description: Production placeholder
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Front-office JWT from /api/front/auth/login or OAuth callback
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    CategorySlugParam:
      name: slug
      in: path
      required: true
      schema:
        type: string
...
    LocaleHeader:
      name: x-vex-locale
      in: header
      description: 前台语言代码；与 site_languages 绑定，同时决定价格币种
      schema:
        type: string
        example: en
    LocaleQueryParam:
      name: locale
      in: query
      description: 可选语言覆盖（优先级低于 x-vex-locale）
      schema:
        type: string
        example: en
    AddressTypeQueryParam:
      name: type
      in: query
      description: 按地址类型筛选
      schema:
        type: string
        enum: [shipping, billing]
    ProductIdParam:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type:
            - array
            - 'null'
          items:
            type: object
            additionalProperties: true
      required: [code, message]
      additionalProperties: false
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, pageSize, total, totalPages]
      additionalProperties: false
    Money:
      type: object
      properties:
        currency:
          type: string
          example: USD
        amount:
          type: number
        formatted:
          type: string
      required: [currency, amount, formatted]
      additionalProperties: false
    ImageAsset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        alt:
          type: string
        width:
          type: integer
        height:
          type: integer
      required: [id, url, alt]
      additionalProperties: false
    Brand:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        logo:
          $ref: '#/components/schemas/ImageAsset'
      required: [id, name, slug]
      additionalProperties: false
...
    ProductCard:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        sku:
          type: string
        shortDescription:
          type: string
        coverImage:
          $ref: '#/components/schemas/ImageAsset'
        price:
          $ref: '#/components/schemas/Money'
        compareAtPrice:
          oneOf:
            - $ref: '#/components/schemas/Money'
            - type: 'null'
        purchaseMode:
          type: string
          enum: [buy, inquiry]
        inStock:
          type: boolean
        brand:
          oneOf:
            - $ref: '#/components/schemas/Brand'
            - type: 'null'
      required: [id, name, slug, sku, purchaseMode, inStock]
      additionalProperties: false
...
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductCard'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        facets:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [items, meta, facets]
      additionalProperties: false
...
    Address:
      type: object
      properties:
        id:
          type: string
          format: uuid
        addressType:
          type: string
          enum: [shipping, billing]
        firstName:
          type: string
        lastName:
          type: string
        company:
          type: string
        phone:
          type: string
        countryCode:
          type: string
        state:
          type: string
        city:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        postalCode:
          type: string
        isDefault:
          type: boolean
      required: [id, addressType, firstName, lastName, countryCode, city, addressLine1, postalCode, isDefault]
      additionalProperties: false
...
    AuthTokenResponse:
      type: object
      properties:
        token:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            firstName:
              type: string
            lastName:
              type: string
            company:
              type:
                - string
                - 'null'
            status:
              type: string
              enum: [active, disabled, pending]
          required: [id, email, firstName, lastName, status]
        redirectPath:
          type: string
        message:
          type: string
      required: [token, user]
      additionalProperties: false
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
      required: [currentPassword, newPassword]
      additionalProperties: false
...
    AuthLoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [email, password]
      additionalProperties: false
    AuthRegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        companyName:
          type: string
        country:
          type: string
        industry:
          type: string
        companySize:
          type: string
        website:
          type: string
        taxId:
          type: string
        termsAccepted:
          type: boolean
        privacyAccepted:
          type: boolean
        exportComplianceAccepted:
          type: boolean
        _quick:
          type: boolean
          description: 快速结账注册；走轻量注册流程
      required: [email, password, firstName, lastName]
      additionalProperties: false
...
paths:
...
  /api/front/categories/{slug}/products:
    get:
      tags: [Categories]
      summary: 获取分类商品列表
      parameters:
        - $ref: '#/components/parameters/CategorySlugParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 分类商品列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '404':
          description: 分类不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/front/addresses:
    get:
      tags: [Addresses]
      summary: 获取当前用户地址列表
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AddressTypeQueryParam'
      responses:
        '200':
          description: 地址列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
    post:
      tags: [Addresses]
      summary: 新增地址
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Address'
                - type: object
                  required: [addressType, firstName, lastName, countryCode, city, addressLine1, postalCode]
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
  /api/front/addresses/{id}:
    patch:
      tags: [Addresses]
      summary: 更新地址
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
      responses:
        '200':
          description: 更新成功
    delete:
      tags: [Addresses]
      summary: 删除地址
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 删除成功
  /api/front/auth/register:
    post:
      tags: [Auth]
      summary: 用户注册（轻量或 B2B 完整注册）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
  /api/front/auth/login:
    post:
      tags: [Auth]
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          description: 登录失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/front/auth/change-password:
    post:
      tags: [Auth]
      summary: 登录后修改密码
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
        '400':
          description: 当前密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/front/compare/items:
    get:
      tags: [Compare]
      summary: 获取登录用户对比列表（匿名用户请使用 localStorage）
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LocaleHeader'
        - $ref: '#/components/parameters/LocaleQueryParam'
      responses:
        '200':
          description: 对比列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  locale:
                    type: string
                  items:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                required: [locale, items]
    post:
      tags: [Compare]
      summary: 添加商品到对比列表
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LocaleHeader'
        - $ref: '#/components/parameters/LocaleQueryParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                  format: uuid
              required: [productId]
      responses:
        '201':
          description: 添加成功
        '409':
          description: 对比列表已达上限（4 个）
  /api/front/compare/items/{productId}:
    delete:
      tags: [Compare]
      summary: 从对比列表移除商品
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      responses:
        '204':
          description: 删除成功
...